Move `lib/utils/prompt` to `api/utils/prompt` (#32334)
* Move /lib/utils/prompt to /api/utils/prompt.

* Replace uses of lib/utils/prompt with api/utils/prompt and delete pacakge.

* go mod tidy.
