+++ lib/client/mfa/prompt.go
@@ -29,10 +29,10 @@ import (
 
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/client/proto"
+	"github.com/gravitational/teleport/api/utils/prompt"
 	wancli "github.com/gravitational/teleport/lib/auth/webauthncli"
 	wantypes "github.com/gravitational/teleport/lib/auth/webauthntypes"
 	"github.com/gravitational/teleport/lib/auth/webauthnwin"
-	"github.com/gravitational/teleport/lib/utils/prompt"
 )
 
 var log = logrus.WithFields(logrus.Fields{
